Move tls config implementation to util file
